pyproject.toml
[tool.poetry]
name = "skitso"
version = "0.1.13"
description = "Ligth Weigth Sketching Tool"
authors = ["Javier Mansilla <javimansilla@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = ">=10.4.0, < 12.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
